Crate reth_eth_wire

Implementation of the eth wire protocol.

Feature Flags

serde (default): Enable serde support

arbitrary: Adds proptest and arbitrary support for wire types.

Re-exports

pub use crate::capability::Capability;

pub use crate::muxdemux::MuxDemuxStream;

pub use crate::muxdemux::StreamClone;

pub use crate::types::EthVersion;

pub use builder::*;

pub use types::*;

Modules

builder

Builder structs for messages.

capability

All capability related types

errors

Error types for stream variants

multiplex

Rlpx protocol multiplexer and satellite stream

muxdemux

MuxDemuxer allows for multiple capability streams to share the same p2p connection. De-/ muxing the connection offers two stream types MuxDemuxStream and StreamClone. MuxDemuxStream is the main stream that wraps the p2p connection, only this stream can advance transfer across the network. One MuxDemuxStream can have many StreamClones, these are weak clones of the stream and depend on advancing the MuxDemuxStream to make progress.

protocol

A Protocol defines a P2P subprotocol in a RLPx connection

types

Types for the eth wire protocol.

Structs

EthStream

An EthStream wraps over any Stream that yields bytes and makes it compatible with eth-networking protocol messages, which get RLP encoded/decoded.

HelloMessage

Raw rlpx protocol message used in the p2p handshake, containing information about the supported RLPx protocol version and capabilities.

HelloMessageBuilder

Builder for HelloMessageWithProtocols

HelloMessageWithProtocols

This is a superset of HelloMessage that provides additional protocol Protocol information about the number of messages used by each capability in order to do proper message ID multiplexing.

P2PStream

A P2PStream wraps over any Stream that yields bytes and makes it compatible with p2p protocol messages.

UnauthedEthStream

An un-authenticated EthStream. This is consumed and returns a EthStream after the Status handshake is completed.

UnauthedP2PStream

An un-authenticated P2PStream. This is consumed and returns a P2PStream after the Hello handshake is completed.

Enums

DisconnectReason

RLPx disconnect reason.

P2PMessage

This represents only the reserved p2p subprotocol messages.

P2PMessageID

Message IDs for p2p subprotocol messages.

ProtocolVersion

RLPx p2p protocol version

Constants

MAX_MESSAGE_SIZE

MAX_MESSAGE_SIZE is the maximum cap on the size of a protocol message.

MAX_RESERVED_MESSAGE_ID

MAX_RESERVED_MESSAGE_ID is the maximum message ID reserved for the p2p subprotocol. If there are any incoming messages with an ID greater than this, they are subprotocol messages.

Traits

CanDisconnect

This trait is meant to allow higher level protocols like eth to disconnect from a peer, using lower-level disconnect functions (such as those that exist in the p2p protocol) if the underlying stream supports it.

DisconnectP2P

Gracefully disconnects the connection by sending a disconnect message and stop reading new messages.