Module reth_eth_wire::muxdemux

MuxDemuxer allows for multiple capability streams to share the same p2p connection. De-/ muxing the connection offers two stream types MuxDemuxStream and StreamClone. MuxDemuxStream is the main stream that wraps the p2p connection, only this stream can advance transfer across the network. One MuxDemuxStream can have many StreamClones, these are weak clones of the stream and depend on advancing the MuxDemuxStream to make progress.

MuxDemuxer filters bytes according to message ID offset. The message ID offset is negotiated upon start of the p2p connection. Bytes received by polling the MuxDemuxStream or a StreamClone are specific to the capability stream wrapping it. When received the message IDs are unmasked so that all message IDs start at 0x0. MuxDemuxStream and StreamClone mask message IDs before sinking bytes to the MuxDemuxer.

For example, EthStream<MuxDemuxStream<P2PStream<S>>> is the main capability stream. Subsequent capability streams clone the p2p connection via EthStream.

When MuxDemuxStream is polled, MuxDemuxer receives bytes from the network. If these bytes belong to the capability stream wrapping the MuxDemuxStream then they are passed up directly. If these bytes however belong to another capability stream, then they are buffered on a channel. When StreamClone is polled, bytes are read from this buffer. Similarly StreamClone buffers egress bytes for MuxDemuxer that are read and sent to the network when MuxDemuxStream is polled.

Structs

MuxDemuxStream

The main stream on top of the p2p stream. Wraps MuxDemuxer and enforces it can’t be dropped before all secondary streams are dropped (stream clones).

MuxDemuxer

Stream MUX/DEMUX acts like a regular stream and sink for the owning stream, and handles bytes belonging to other streams over their respective channels.

StreamClone

More or less a weak clone of the stream wrapped in MuxDemuxer but the bytes belonging to other capabilities have been filtered out.