Module reth_eth_wire::types

Types for the eth wire protocol.

Re-exports

pub use version::EthVersion;

pub use message::EthMessage;

pub use message::EthMessageID;

pub use message::ProtocolMessage;

pub use blocks::*;

pub use broadcast::*;

pub use transactions::*;

pub use state::*;

pub use receipts::*;

Modules

blocks

Implements the GetBlockHeaders, GetBlockBodies, BlockHeaders, and BlockBodies message types.

broadcast

Types for broadcasting new data.

message

receipts

Implements the GetReceipts and Receipts message types.

state

Implements the GetNodeData and NodeData message types.

transactions

Implements the GetPooledTransactions and PooledTransactions message types.

version

Support for representing the version of the eth. Capability and Protocol.

Structs

Status

The status message is used in the eth protocol handshake to ensure that peers are on the same network and are following the same fork.