Module reth_eth_wire::types::broadcast

Types for broadcasting new data.

Structs

BlockHashNumber

A block hash and a block number.

NewBlock

A new block with the current total difficulty, which includes the difficulty of the returned block.

NewBlockHashes

This informs peers of new blocks that have appeared on the network.

NewPooledTransactionHashes66

This informs peers of transaction hashes for transactions that have appeared on the network, but have not been included in a block.

NewPooledTransactionHashes68

Same as NewPooledTransactionHashes66 but extends that that beside the transaction hashes, the node sends the transaction types and their sizes (as defined in EIP-2718) as well.

SharedTransactions

Same as Transactions but this is intended as egress message send from local to many peers.

Transactions

This informs peers of transactions that have appeared on the network and are not yet included in a block.

Enums

NewPooledTransactionHashes

A wrapper type for all different new pooled transaction types