Trait reth_eth_wire::types::broadcast::HandleAnnouncement

pub trait HandleAnnouncement {
    // Required methods
    fn is_empty(&self) -> bool;
    fn len(&self) -> usize;
    fn retain_by_hash(&mut self, f: impl FnMut(TxHash) -> bool);
    fn msg_version(&self) -> EthVersion;
}

Interface for handling announcement data in filters in the transaction manager and transaction pool. Note: this trait may disappear when distinction between eth66 and eth68 hashes is more clearly defined, see https://github.com/paradigmxyz/reth/issues/6148.

Required Methods

fn is_empty(&self) -> bool

The announcement contains no entries.

fn len(&self) -> usize

Returns the number of entries.

fn retain_by_hash(&mut self, f: impl FnMut(TxHash) -> bool)

Retain only entries for which the hash in the entry, satisfies a given predicate.

fn msg_version(&self) -> EthVersion

Returns the announcement version, either EthVersion::Eth66 or EthVersion::Eth68.

Object Safety

This trait is not object safe.