Module reth_eth_wire::types::message

Structs

ProtocolBroadcastMessage

Represents messages that can be sent to multiple peers.

ProtocolMessage

An eth protocol message, containing a message ID and payload.

RequestPair

This is used for all request-response style eth protocol messages. This can represent either a request or a response, since both include a message payload and request id.

Enums

EthBroadcastMessage

Represents broadcast messages of EthMessage with the same object that can be sent to multiple peers.

EthMessage

Represents a message in the eth wire protocol, versions 66, 67 and 68.

EthMessageID

Represents message IDs for eth protocol messages.

Constants

MAX_MESSAGE_SIZE

MAX_MESSAGE_SIZE is the maximum cap on the size of a protocol message.