Trait reth_eth_wire::CanDisconnect

pub trait CanDisconnect<T>: Sink<T> + Unpin {
    // Required method
    fn disconnect<'life0, 'async_trait>(
        &'life0 mut self,
        reason: DisconnectReason
    ) -> Pin<Box<dyn Future<Output = Result<(), <Self as Sink<T>>::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

This trait is meant to allow higher level protocols like eth to disconnect from a peer, using lower-level disconnect functions (such as those that exist in the p2p protocol) if the underlying stream supports it.

Required Methods

fn disconnect<'life0, 'async_trait>( &'life0 mut self, reason: DisconnectReason ) -> Pin<Box<dyn Future<Output = Result<(), <Self as Sink<T>>::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Disconnects from the underlying stream, using a DisconnectReason as disconnect information if the stream implements a protocol that can carry the additional disconnect metadata.